Rename `rbs_location.*` and `rbs_string.*` to `location.*` and `string.*` (#34)
This aligns with the naming convention of Prism:

- Private headers are prefixed with `rbs_` and placed under
`include/rbs/util/`.
- Public headers are placed under `include/rbs/` without the `rbs_`
prefix.
