+++ ext/rbs_extension/legacy_location.h
