+++ include/rbs/string.h
